Calculate Running Totals
The cumulative sums of all transactions carried out in December 2020 need to be compiled in a report that is aggregated based on the card type.

Initially, the code selects a specific date to avoid the presence of multiple transactions on the same day by the same card type, or even the same card number. Following that, the column "card_type_name" is used. The last column in the query is "transaction_running_total." To determine the running total, a window function is used. The column "amount" is summed first. Then, the OVER() clause is used to specify that the running total should be calculated on a card type level; thus, the clause reads PARTITION BY cty.card_type_name. Finally, the running total is calculated from the earliest to the latest date using ORDER BY date ASC.

To retrieve the data, three tables must be joined. The first join connects the tables "card_transaction" and "card_number." The second join refers to the table "card_type," which is how the card type name is obtained in the report. Aliases have been assigned to all joined tables, resulting in fewer keystrokes. The results are filtered using the WHERE clause, which only retrieves transactions that occurred in December 2020. Data has been sorted in ascending order based on the card type name.